Fix kit root note int definition
Should be u8 instead i32 because it's being stored as a default in flash storage in a u8 buffer.

+++ src/deluge/io/midi/midi_engine.h
@@ -64,7 +64,7 @@ public:
 	bool midiThru;
 	bool midiFollow;
 	LearnedMIDI midiFollowChannelType[kNumMIDIFollowChannelTypes]; //Synth, Kit, Param
-	int32_t midiFollowKitRootNote;
+	uint8_t midiFollowKitRootNote;
 	bool midiFollowDisplayParam;
 	bool midiFollowFeedback;
 	MIDIFollowFeedbackAutomationMode midiFollowFeedbackAutomation;
